+++ estore-api/src/main/java/com/estore/api/estoreapi/model/Product.java
@@ -26,9 +26,9 @@ public class Product {
     @JsonProperty("type")
     private String type;
     @JsonProperty("modPrice")
-    private float modPrice;
+    private double modPrice;
     @JsonProperty("ingredients")
-    private Map<String, Float> ingredients;
+    private Map<String, Double> ingredients;
 
     /**
      * Create a product with the given id and name
@@ -48,7 +48,7 @@ public class Product {
      *                 value, i.e. 0 for int
      */
     public Product(@JsonProperty("id") int id, @JsonProperty("name") String name, @JsonProperty("type") String type,
-            @JsonProperty("modPrice") float modPrice, @JsonProperty("ingredients") Map<String, Float> ingredients) {
+            @JsonProperty("modPrice") double modPrice, @JsonProperty("ingredients") Map<String, Double> ingredients) {
         this.id = id;
         this.name = name;
         this.type = type;
@@ -109,7 +109,7 @@ public class Product {
      * 
      * @param modPrice The modified price of the product
      */
-    public void setModPrice(float modPrice) {
+    public void setModPrice(double modPrice) {
         this.modPrice = modPrice;
     }
 
@@ -118,7 +118,7 @@ public class Product {
      * 
      * @return The modified price of the product
      */
-    public float getModPrice() {
+    public double getModPrice() {
         return modPrice;
     }
 
@@ -128,7 +128,7 @@ public class Product {
      * 
      * @param ingredients The ingredients of the product
      */
-    public void setIngredients(Map<String, Float> ingredients) {
+    public void setIngredients(Map<String, Double> ingredients) {
         this.ingredients = ingredients;
     }
 
@@ -137,7 +137,7 @@ public class Product {
      * 
      * @return The ingredients of the product
      */
-    public Map<String, Float> getIngredients() {
+    public Map<String, Double> getIngredients() {
         return ingredients;
     }
